CMP 변환 작업(4)
