+++ MTManger/composeApp/src/commonMain/kotlin/com/wonddak/mtmanger/room/dao/CategoryListDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.wonddak.mtmanger.room.categoryList
+import com.wonddak.mtmanger.room.entity.categoryList
 import kotlinx.coroutines.flow.Flow
 
 @Dao
