+++ MTManger/composeApp/src/commonMain/kotlin/com/wonddak/mtmanger/room/entity/categoryList.kt
@@ -1,0 +1,11 @@
+package com.wonddak.mtmanger.room.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class categoryList(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int?,
+    val name: String
+)
