+++ MTManger/composeApp/src/commonMain/kotlin/com/wonddak/mtmanger/ui/view/dialog/ContactButton.kt
@@ -2,9 +2,7 @@ package com.wonddak.mtmanger.ui.view.dialog
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.wonddak.mtmanger.room.SimplePerson
-import com.wonddak.mtmanger.viewModel.MTViewModel
-import org.koin.compose.koinInject
+import com.wonddak.mtmanger.room.entity.*
 
 @Composable
 internal expect fun ContactButton(
