+++ MTManger/composeApp/src/commonMain/kotlin/com/wonddak/mtmanger/ui/view/home/PersonView.kt
@@ -36,8 +36,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.wonddak.mtmanger.model.Resource
-import com.wonddak.mtmanger.room.MtDataList
-import com.wonddak.mtmanger.room.Person
+import com.wonddak.mtmanger.room.entity.*
 import com.wonddak.mtmanger.toPriceString
 import com.wonddak.mtmanger.ui.theme.maple
 import com.wonddak.mtmanger.ui.theme.match1
